+++ noodles-cram/src/lib.rs
@@ -21,6 +21,11 @@ pub use self::{
 };
 
 #[cfg(feature = "async")]
-pub use self::r#async::io::{Reader as AsyncReader, Writer as AsyncWriter};
+#[deprecated(since = "0.69.0", note = "Use `cram::r#async::io::Reader` instead.")]
+pub use self::r#async::io::Reader as AsyncReader;
+
+#[cfg(feature = "async")]
+#[deprecated(since = "0.69.0", note = "Use `cram::r#async::io::Writer` instead.")]
+pub use self::r#async::io::Writer as AsyncWriter;
 
 static MAGIC_NUMBER: &[u8] = b"CRAM";